Fixed bounds for the check function
When we changed our logic, the bounds of the divisor are different. N can be more than half of the number, because we are no longer summing the digits N, as it now represents the divisor. In other words, N is how many digits are summed to get the target.